refactor(client-js): align a11y file names
