+++ client-js/rtvi/index.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-export * from "./a11y-walker";
+export * from "./a11y_walker";
 export * from "./common_types";
 export * from "./errors";
 export * from "./events";
