+++ client-js/tests/a11y_walker.spec.ts
@@ -6,7 +6,7 @@
 
 import { beforeEach, describe, expect, it } from "@jest/globals";
 
-import { findElementByRef, snapshotDocument } from "../rtvi/a11y-walker";
+import { findElementByRef, snapshotDocument } from "../rtvi/a11y_walker";
 
 function html(body: string): HTMLElement {
   document.body.innerHTML = body;
